requirements.txt
time
requests
secrets
user_agent
python-telegram-bot
random